injury-lens-ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
Pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
torch
transformers